rename kbn-utility-types/test --> kbn-utility-types/jest
