+++ packages/kbn-utility-types/jest/package.json
@@ -1,0 +1,3 @@
+{
+  "types": "../target/jest/index.d.ts"
+}
