--- packages/kbn-utility-types/test/package.json
@@ -1,4 +1,0 @@
-{
-  "types": "../target/test/index.d.ts"
-}
-
