+++ src/core/public/apm_system.test.ts
@@ -18,9 +18,8 @@
  */
 
 jest.mock('@elastic/apm-rum');
-import type { DeeplyMockedKeys } from '@kbn/utility-types/test';
+import type { DeeplyMockedKeys } from '@kbn/utility-types/jest';
 import { init, apm } from '@elastic/apm-rum';
-import { DeeplyMockedKeys } from '../typings';
 import { ApmSystem } from './apm_system';
 
 const initMock = init as jest.Mocked<typeof init>;
