+++ src/core/public/chrome/chrome_service.mock.ts
@@ -18,7 +18,7 @@
  */
 import { BehaviorSubject } from 'rxjs';
 import type { PublicMethodsOf } from '@kbn/utility-types';
-import type { DeeplyMockedKeys } from '@kbn/utility-types/test';
+import type { DeeplyMockedKeys } from '@kbn/utility-types/jest';
 import { ChromeBadge, ChromeBrand, ChromeBreadcrumb, ChromeService, InternalChromeStart } from './';
 
 const createStartContractMock = () => {
