+++ src/core/public/notifications/notifications_service.mock.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 import type { PublicMethodsOf } from '@kbn/utility-types';
-import type { MockedKeys } from '@kbn/utility-types/test';
+import type { MockedKeys } from '@kbn/utility-types/jest';
 import {
   NotificationsService,
   NotificationsSetup,
