+++ src/core/server/core_context.mock.ts
@@ -18,7 +18,7 @@
  */
 
 import { REPO_ROOT } from '@kbn/dev-utils';
-import type { DeeplyMockedKeys } from '@kbn/utility-types/test';
+import type { DeeplyMockedKeys } from '@kbn/utility-types/jest';
 import { CoreContext } from './core_context';
 import { Env, IConfigService } from './config';
 import { configServiceMock, getEnvOptions } from './config/mocks';
