+++ src/core/server/elasticsearch/client/mocks.ts
@@ -18,7 +18,7 @@
  */
 import { Client, ApiResponse } from '@elastic/elasticsearch';
 import { TransportRequestPromise } from '@elastic/elasticsearch/lib/Transport';
-import type { DeeplyMockedKeys } from '@kbn/utility-types/test';
+import type { DeeplyMockedKeys } from '@kbn/utility-types/jest';
 import { ElasticsearchClient } from './types';
 import { ICustomClusterClient } from './cluster_client';
