+++ src/plugins/data/common/search/search_source/mocks.ts
@@ -18,7 +18,7 @@
  */
 
 import { BehaviorSubject } from 'rxjs';
-import type { MockedKeys } from '@kbn/utility-types/test';
+import type { MockedKeys } from '@kbn/utility-types/jest';
 import { uiSettingsServiceMock } from '../../../../../core/public/mocks';
 
 import { SearchSource } from './search_source';
