+++ src/plugins/data/public/search/es_search/get_es_preference.test.ts
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import type { MockedKeys } from '@kbn/utility-types/test';
+import type { MockedKeys } from '@kbn/utility-types/jest';
 import { getEsPreference } from './get_es_preference';
 import { CoreStart } from '../../../../../core/public';
 import { coreMock } from '../../../../../core/public/mocks';
