+++ src/plugins/security_oss/public/plugin.mock.ts
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import type { DeeplyMockedKeys } from '@kbn/utility-types/test';
+import type { DeeplyMockedKeys } from '@kbn/utility-types/jest';
 import { mockInsecureClusterService } from './insecure_cluster_service/insecure_cluster_service.mock';
 import { SecurityOssPluginSetup, SecurityOssPluginStart } from './plugin';
